requirements.txt
asgiref==3.8.1
Django==5.0.3
django-model-utils==4.4.0
djangorestframework==3.15.1
djoser==2.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
sentry-sdk==1.4.1
sqlparse==0.4.4
typing_extensions==4.10.0
